pyproject.toml
[project]
name = "acrophase-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "favicon>=0.7.0",
    "httpx>=0.28.1",
    "make>=0.1.6.post2",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
